It is time for refactoring. I have done some refactoring in LifeDecisionRule class and it's Test class.
In Test class, multiple lines were repeating. As per DRY principle, we shouldn't repeat ourself. So I have extracted multiple methods for the same.

Another refactoring done in LifeDecisionRule class. I see for each request we are calculating neighbours everytime, this is actually not required. I have refactor this part and now, it will be created during Grid creation time so that we don't have to repeat ourself.

Another refactoring is done at cell level, instead of playing simple boolean, created an Object of type Cell. It will help to manage neighbours and whether cell is active or not.

I have created cell class as an abstract class and its sub classes Alive and Dead cell. This will help to give a more meaning to a cell.