Another refactoring is done after previous one. I am thinking now that Neighbour class is not required any more after having neighbours defined at cell level. Also, refactor grid class to add this neighbours.
In LifeDecisionTest class, also few of the test cases are not required any more. But we need to add those test cases in GridTest.

I also observed that, when we are making a cell active/dead we are creating a new object and setting up neighbours again. But problem is, all the neighbours have different reference of the object. So either we have to update reference for each of the neighbours or instead of AliveCell/DeadCell use enum status Alive/Dead. This way we need to update just the status without need to update the neighbours. So I have removed this 2 classes.